+++ src/shared/assets/icons/ErrorIcon.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const ErrorIcon = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <circle cx="12" cy="12" r="10" />
+    <line x1="8" y1="8" x2="16" y2="16" />
+    <line x1="16" y1="8" x2="8" y2="16" />
+  </svg>
+);
+
+export default ErrorIcon;
