+++ src/shared/assets/icons/InfoIcon.tsx
@@ -1,0 +1,18 @@
+const InfoIcon = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <circle cx="12" cy="12" r="10" />
+    <circle cx="12" cy="8" r="1.2" fill="currentColor" stroke="none" />
+    <line x1="12" y1="11" x2="12" y2="17" />
+  </svg>
+);
+
+export default InfoIcon;
