+++ src/shared/assets/icons/SuccessIcon.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+const SuccessIcon = () => (
+  <svg
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <circle cx="12" cy="12" r="10" />
+    <path d="M8 12.5 L11 15.5 L16 9.5" />
+  </svg>
+);
+
+export default SuccessIcon;
