+++ src/shared/assets/icons/index.ts
@@ -1,1 +1,4 @@
 export { default as ChevronDown } from "./ChevronDown";
+export { default as InfoIcon } from "./InfoIcon";
+export { default as ErrorIcon } from "./ErrorIcon";
+export { default as SuccessIcon } from "./SuccessIcon";
